pyproject.toml
[project]
name = "sakuraisenrin"
version = "3.0.0"
description = "🥰 A powerful and cute QQ bot designed for galgame enthusiasts, powered by NoneBot2. ｜ 原来你也玩 GalGame？"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "arrow>=1.4.0",
    "greenlet>=3.3.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot2[fastapi]>=2.4.4",
    "pil-utils>=0.2.2",
    "sentry-sdk>=2.50.0",
    "sqlalchemy>=2.0.46",
]


[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"


[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "ANN",   # flake8-annotations
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "TID",   # flake8-tidy-imports
    "RUF",   # Ruff-specific rules
    "COM",   # flake8-commas
]
ignore = [
    "ANN401",
    "E402",   # module-import-not-at-top-of-file
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
